refactor(xtts): use build_hf_gpt_transformer from tortoise
